requirements.txt
opencv-python
pillow
pandas
numpy
openpyxl
